insightface/requirements.txt
absl-py==2.1.0
addict==2.4.0
albucore==0.0.21
albumentations==1.4.23
annotated-types==0.7.0
asttokens==3.0.0
astunparse==1.6.3
attrs==24.2.0
autocommand==2.2.2
backports.tarfile==1.2.0
basicsr==1.4.2
batch-face==1.5.0
beautifulsoup4==4.12.3
blinker==1.9.0
brisque==0.0.16
certifi==2024.8.30
cffi==1.17.1
cgen==2020.1
charset-normalizer==3.4.0
click==8.1.7
codepy==2023.1
colorama==0.4.6
coloredlogs==15.0.1
contourpy==1.3.1
cycler==0.12.1
Cython==3.0.11
decorator==4.4.2
deepface==0.0.93
defusedxml==0.7.1
dlib==19.24.6
easydict==1.13
eval_type_backport==0.2.0
exceptiongroup==1.2.2
executing==2.1.0
face-recognition==1.3.0
face_recognition_models==0.3.0
facenet-pytorch==2.6.0
facexlib==0.3.0
faiss-cpu==1.9.0.post1
fer==22.5.1
ffmpeg==1.4
filelock==3.16.1
filterpy==1.4.5
fire==0.7.0
Flask==3.1.0
Flask-Cors==5.0.0
flatbuffers==24.3.25
fonttools==4.55.0
fsspec==2024.10.0
future==1.0.0
gast==0.6.0
gdown==5.2.0
genpy==2022.1
gitdb==4.0.11
GitPython==3.1.43
google-pasta==0.2.0
grpcio==1.68.0
gunicorn==23.0.0
h5py==3.12.1
huggingface-hub==0.26.5
humanfriendly==10.0
idna==3.10
imageio==2.36.1
imageio-ffmpeg==0.5.1
immutabledict==4.2.1
immutables==0.21
importlib_metadata==8.0.0
importlib_resources==6.4.0
inflect==7.3.1
insightface==0.7.3
ipython==8.30.0
islpy==2024.2
itsdangerous==2.2.0
jaraco.collections==5.1.0
jaraco.context==5.3.0
jaraco.functools==4.0.1
jaraco.text==3.12.1
jax==0.4.35
jaxlib==0.4.35
jedi==0.19.2
Jinja2==3.1.4
joblib==1.4.2
keras==3.7.0
kiwisolver==1.4.7
lazy_loader==0.4
libclang==18.1.1
libsvm-official==3.35.0
llvmlite==0.43.0
lmdb==1.5.1
loopy==2024.1
lxml==5.3.0
lz4==4.3.3
Mako==1.3.6
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.9.2
matplotlib-inline==0.1.7
mdurl==0.1.2
mediapipe==0.10.18
ml-dtypes==0.4.1
more-itertools==10.3.0
moviepy==1.0.3
mpmath==1.3.0
mtcnn==1.0.0
namex==0.0.8
networkx==3.4.2
numba==0.60.0
numpy==1.26.4
onnx==1.17.0
onnxruntime==1.20.1
opencv-contrib-python==4.10.0.84
opencv-python==4.10.0.84
opencv-python-headless==4.10.0.84
opencv-transforms==0.0.6
opt_einsum==3.4.0
optree==0.13.1
packaging==24.2
pandas==2.2.3
parso==0.8.4
pexpect==4.9.0
pillow==10.2.0
platformdirs==4.3.6
prettytable==3.12.0
proglog==0.1.10
prompt_toolkit==3.0.48
protobuf==4.25.5
psutil==6.1.0
ptyprocess==0.7.0
pure_eval==0.2.3
py-cpuinfo==9.0.0
pycocotools==2.0.8
pycparser==2.22
pydantic==2.10.2
pydantic_core==2.27.1
Pygments==2.18.0
pymbolic==2024.2
pyparsing==3.2.0
pyrsistent==0.20.0
PySocks==1.7.1
pytato==2021.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytools==2024.1.17
pytz==2024.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
retina-face==0.0.17
rich==13.9.4
safetensors==0.4.5
scikit-image==0.24.0
scikit-learn==1.6.0
scipy==1.14.1
seaborn==0.13.2
sentencepiece==0.2.0
simsimd==6.2.1
six==1.16.0
sixdrepnet==0.1.6
smmap==5.0.1
sounddevice==0.5.1
soupsieve==2.6
stack-data==0.6.3
stringzilla==3.10.10
supervision==0.25.0
sympy==1.13.1
tb-nightly==2.19.0a20241121
tensorboard==2.18.0
tensorboard-data-server==0.7.2
tensorflow==2.18.0
tensorflow-io-gcs-filesystem==0.37.1
termcolor==2.5.0
tf_keras==2.18.0
thop==0.1.1.post2209072238
threadpoolctl==3.5.0
tifffile==2024.9.20
tokenizers==0.21.0
tomli==2.2.1
torch==2.5.1
torchvision==0.20.1
tqdm==4.67.1
traitlets==5.14.3
transformers==4.47.0
typeguard==4.3.0
typing_extensions==4.12.2
tzdata==2024.2
ultralytics==8.3.48
ultralytics-thop==2.0.13
urllib3==2.2.3
useful_types==0.2.1
wcwidth==0.2.13
Werkzeug==3.1.3
wrapt==1.17.0
yapf==0.43.0
zipp==3.19.2
